--- Interfaces/breadcrumb.ts
@@ -1,8 +1,0 @@
-export interface BreadcrumbItem {
-    link: string
-    title: string
-}
-
-export interface BreadcrumbProps {
-    BreadcrumbList: BreadcrumbItem[]
-}
