+++ src/pages/users/DeleteUser.tsx
@@ -3,7 +3,7 @@ import { Navigate, useParams } from "react-router-dom";
 import { ONSButton, ONSPanel } from "blaise-design-system-react-components";
 import { deleteUser } from "../../utilities/http";
 import { ReturnPanel, UserRouteParams } from "../../../interfaces/usersPage";
-import { BreadcrumbItem } from "../../../interfaces/breadcrumb";
+import { BreadcrumbItem } from "../../../interfaces/";
 import Breadcrumbs from "../../Components/Breadcrumbs";
 
 function DeleteUser(): ReactElement {
